Move referenceLoader
