+++ apps/server/src/modules/authorization/authorization.service.spec.ts
@@ -7,7 +7,7 @@ import { AuthorizationContextBuilder } from './authorization-context.builder';
 import { AuthorizationHelper } from './authorization.helper';
 import { AuthorizationService } from './authorization.service';
 import { ForbiddenLoggableException } from './errors/forbidden.loggable-exception';
-import { ReferenceLoader } from './reference.loader';
+import { ReferenceLoader } from './domain/reference/reference.loader';
 import { RuleManager } from './rule-manager';
 import { AuthorizableReferenceType, Rule } from './types';
