+++ apps/server/src/modules/authorization/authorization.service.ts
@@ -4,7 +4,7 @@ import { AuthorizableObject } from '@shared/domain/domain-object';
 import { ErrorUtils } from '@src/core/error/utils';
 import { AuthorizationHelper } from './authorization.helper';
 import { ForbiddenLoggableException } from './errors/forbidden.loggable-exception';
-import { ReferenceLoader } from './reference.loader';
+import { ReferenceLoader } from './domain/reference/reference.loader';
 import { RuleManager } from './rule-manager';
 import { AuthorizableReferenceType, AuthorizationContext } from './types';
