+++ apps/server/src/modules/authorization/domain/reference/reference.loader.ts
@@ -14,7 +14,7 @@ import {
 } from '@shared/repo';
 import { BoardDoAuthorizableService } from '@src/modules/board/service';
 import { ContextExternalToolAuthorizableService } from '@src/modules/tool/context-external-tool/service';
-import { AuthorizableReferenceType } from './types';
+import { AuthorizableReferenceType } from '../../types';
 
 // replace later with general "base" do-repo
 type RepoType =
